Fix equality checks
Fixes two equality check issues

1. ACL whitelist and blacklist attributes can be strings or array of
strings. Whitelist from the server comes accross as an array of strings,
config is parsed as a string. This means the diff function would always
detect a diff, even if there was no diff

2. Transform request/response always showed up as needing an update.
This is because the response from the server included empty objects for
the options for both plugins (think 'body': {} or 'json': {}). Any
object that has a key that is empty is now removed from the server
object before the diff comparison.

There may be other equality check issues I didn't find. The two above
are the only two I found running kong_sync dry against dev and prod.