Fix integration tests (#9)
* Make integration tests runnable

* Run integration tests serially

* Fix add SNI request

Kong 0.14 changed the format of the add SNI payload from
{ "name": "example.com", "ssl_certificate_id": "21b69eab-09d9-40f9-a55e-c4ee47fada68" }
to
{ "name": "example.com", "certificate": { "id": "21b69eab-09d9-40f9-a55e-c4ee47fada68" } }

* Delete redundant integration tests

These tests are the same as the "config" tests, but they include the
config inline as a JS object. The config tests are preferable because
they also test the yaml format, which we use.

* Increase jest timeout

Kongfig retries failed requests a few times, which takes longer than
the default 5 second timeout. When that happens, the test output hides
the actual error messages from kong and instead shows a timeout error.

* Include kong host in npm script

* Prevent error on null log message

* Fix second sni payload bug

* Update snapshots and delete obsolete ones

Removes all of the snapshots associated with removed tests.
Changes within the snapshots are the result of the SNI object changes
and the addition of the consumer.custom_id field

* Add compatibility notice to readme

* Run pretest script before integration tests

npm automatically runs the pretest script before the test script, so
this explicitly ensures the same thing for the integration tests.

* Revert "Run pretest script before integration tests"

This reverts commit c333f5d01d0145f13900ec618cd3c6f31bd55788.

* Import from uncompiled files in tests for easier debugging