Use route names (#26)
* Include route names in kong api requests

Since kong 1.0 introduced route names, kongfig's custom route name
handling doesn't play nicely with the official attribute. This commit
lets kongfig use the new route names while still allowing the old format
to work.

* Ensure attributes.name is always set

* Add some tests for route name changes

* Revert "Ensure attributes.name is always set"

This reverts commit b58511dfbedbe38c3ae8d3ff3597072e10b05603.

* Remove useless tests

* Use route names as primary identifier

* Update route integration test

These tests (updating route names) no longer make sense: now that names
are the primary identifier, kongfig can't update them. For the same
reason, a route without a name (the previous starting point of the
tests) is no longer a valid state for kongfig.

* Error if routes don't have names

* Add better route test

* Allow both locations of route name in validation

* Test that error is thrown when route has no name

* Add more tests and stop supporting attributes.name for routes

This matches how kongfig treats other entities (like services) that
have a name property which can be used as a primary key. Supporting both
makes the code overly complicated and is inconsistent.

* Fix missing name test

* Remove stray console log

* Remove unneeded name to id translation

* Update snapshots

There shouldn't have been a change to the snapshot from this work, but
there was because the snapshots contained a bug. Previously the route
was being recreated (because it had no id) during the tests – now that
request is accurately snapshotted as a no-op.

* Remove logging statements

* 3.0.0